Actualiza datos semilla y mejora estructura de servicios
Se han actualizado los datos semilla en `ConsultorioContext` para incluir una nueva modalidad de pago "Transferencia" y se han modificado los honorarios y notas de las sesiones.

En `ConsultorioContextModelSnapshot.cs`, se han ajustado las fechas de las deudas y se han cambiado los tipos de datos de `Monto` y `Honorarios` a `decimal(65,30)`.

Se ha eliminado el archivo `WeatherForecast.cs` y se ha comentado una cadena de conexión para un servidor remoto en `appsettings.json`.

Se ha añadido el paquete `FontAwesome.Sharp` en `Desktop.csproj` para utilizar iconos en la aplicación.

Se han creado nuevas clases para la vista `MenuPrincipalView`, reemplazando a `Form1`, y se han implementado nuevas clases y métodos en `ApiEndpoints.cs` y `IGenericService.cs` para manejar los endpoints de la API y operaciones CRUD.

Además, se han actualizado los archivos de recursos para incluir una URL de API local y se ha implementado la clase `GenericService<T>` para mejorar la gestión de servicios utilizando `HttpClient`.

+++ Service/Models/Pago.cs
@@ -20,7 +20,7 @@ namespace Service.Models
         public virtual ModalidadPago? ModalidadPago { get; set; }
 
         [Required(ErrorMessage = "El campo Monto es obligatorio.")]
-        public int Monto { get; set; } 
+        public decimal Monto { get; set; } 
 
         [Required(ErrorMessage = "El campo Fecha es obligatorio.")]
         [DataType(DataType.Date)]
